completed switch to cartesian coordinates
phased out linear coordinates in favour of cartesian coordinates
everywhere. The project is now back up to where it was before the
change.

+++ src/board.rs
@@ -1,1 +1,2 @@
-pub const BOARD_SIZE: [i8; 2] = [8, 8];
+// Board size is able to be changed, but unit tests will break
+pub const BOARD_SIZE: [usize; 2] = [8, 8];
